modifica estructura del proyecto
