+++ src/js/app/common/components/RemovableListItem.js
