+++ src/js/app/estudios/estudioReducer.js
@@ -1,5 +1,8 @@
 export function estudioReducer(state=[], action){
     switch(action.type){
+        case 'FETCH_ESTUDIOS_DIARIOS':
+        case 'CANCEL_ESTUDIOS_DIARIOS':
+            return [];
         case 'LOAD_ESTUDIOS_DIARIOS':
             return action.data;
         default:
