+++ src/js/app/layout/components/Header.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Dropdown } from 'react-bootstrap';
-import { smoothlyMenu } from '../../layout/Helpers';
+import { smoothlyMenu } from '../Helpers';
 
 class Header extends React.Component {
