+++ src/js/app/reducers.js
@@ -2,11 +2,11 @@ import { combineReducers } from 'redux';
 import { routerReducer } from 'react-router-redux';
 import { combineEpics } from 'redux-observable';
 
-import { userReducer } from './reducers/userReducer';
-import { estudioReducer } from './reducers/estudioReducer';
+import { userReducer } from './layout/userReducer';
+import { estudioReducer } from './estudios/estudioReducer';
 
-import { userEpic } from './epics/userEpic';
-import { estudioEpic } from './epics/estudioEpic';
+import { userEpic } from './layout/userEpic';
+import { estudioEpic } from './estudios/estudioEpic';
 
 export const rootEpic = combineEpics(
   userEpic,
