--- src/js/app/reducers/userReducer.js
@@ -1,10 +1,0 @@
-export function userReducer(state={}, action){
-    switch(action.type){
-        case 'USER_LOGIN':
-            return action.user;
-        case 'USER_LOGOUT':
-            return {};
-        default:
-            return state;
-    }
-}
