+++ src/js/routes.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import { Route, IndexRoute } from 'react-router';
-import App from './app/App';
-import HomePage from './app/components/pages/HomePage';
-import EstudiosPage from './app/components/pages/EstudiosPage';
+import Layout from './app/layout/Layout';
+import HomePage from './app/home/HomePage';
+import EstudiosPage from './app/estudios/EstudiosPage';
 
 export default (
-    <Route path="/" component={App}>
+    <Route path="/" component={Layout}>
         <IndexRoute component={HomePage}/>
         <Route path="/estudios" component={EstudiosPage}/>
     </Route>
